requirements.txt
fastmcp==2.13.0
mcp>=1.19.0
google-genai==1.47.0
python-dotenv==1.2.1
